reasoning/requirements.txt
# EcoRevive Reasoning Engine Dependencies
# Layer 2: Gemini-powered restoration intelligence

# Core Gemini SDK
google-genai>=1.0.0

# Environment variables
python-dotenv>=1.0.0

# Image processing
Pillow>=10.0.0

# Numerical computing (already installed for Layer 1)
numpy>=1.24.0

# Visualization (for severity map coloring)
matplotlib>=3.7.0

# PDF generation
reportlab>=4.0.0
